+++ Models/Reminder/ReminderRecord.cs
@@ -5,7 +5,7 @@
         public int Id { get; set; }
         public int VehicleId { get; set; }
         public DateTime Date { get; set; }
-        public int Mileage { get; set; }
+        public decimal Mileage { get; set; }
         public string Description { get; set; }
         public string Notes { get; set; }
         public bool IsRecurring { get; set; } = false;
