+++ Models/Reminder/ReminderRecordInput.cs
@@ -5,7 +5,7 @@
         public int Id { get; set; }
         public int VehicleId { get; set; }
         public string Date { get; set; } = DateTime.Now.AddDays(1).ToShortDateString();
-        public int Mileage { get; set; }
+        public decimal Mileage { get; set; }
         public string Description { get; set; }
         public string Notes { get; set; }
         public bool IsRecurring { get; set; } = false;
