+++ Models/Report/CostForVehicleByMonth.cs
@@ -5,6 +5,6 @@
         public int MonthId { get; set; }
         public string MonthName { get; set; }
         public decimal Cost { get; set; }
-        public int DistanceTraveled { get; set; }
+        public decimal DistanceTraveled { get; set; }
     }
 }
